ci: Use a GitHub action to validate PR title and description
We established Angular conventional commits in our projects and create
automated changelogs based on the squash commit merges.

This change adds a GitHub action that checks
that the PR title conforms to the Angular
conventional commit style.

Furthermore, the action validates, that the PR description is not empty.

Title validation is skipped, if the PR is marked with the label `bot`,
`ignore-semantic-pull-request` or `rebase`.